requirements.txt
paddlepaddle-gpu==2.4.1.post117
torchvision
torch
pillow
matplotlib